ops/storage.py: Refactor Storage into its own file.
SQLiteStorage and now JujuStorage are put into a file together.
